adding DeferredRenderer and RenderSystem

+++ project-griffin/project-griffin/source/entity/components.h
@@ -28,12 +28,12 @@ COMPONENT_LIST(
 typedef std::bitset<ComponentTypeCount> ComponentMask;
 
 COMPONENT(Position,
-	(glm::vec3, (position, (FieldType::vec3_T, ("3D position coordinate", NIL))))
+	(glm::dvec3, (position, (FieldType::vec3_T, ("3D position coordinate", NIL))))
 )
 
 COMPONENT(Transform,
-	(glm::vec3, (position,    (FieldType::vec3_T, ("3D position coordinate", NIL)))),
-	(glm::quat, (orientation, (FieldType::quat_T, ("Orientation quaternion", NIL))))
+	(glm::dvec3, (position,    (FieldType::vec3_T, ("3D position coordinate", NIL)))),
+	(glm::quat,  (orientation, (FieldType::quat_T, ("Orientation quaternion", NIL))))
 )
 
 COMPONENT(Person,
